split persist_files_handling from otaclient_common.common
